server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "pandas>=2.2.3",
    "numpy>=2.1.2",
    "scikit-learn>=1.5.2",
    "lightgbm",
    "mlflow>=2.22",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "matplotlib>=3.9.2",
    "seaborn>=0.13.2",
    "plotly>=5.24.1",
    "joblib>=1.4.2",
    "colorlog>=6.9.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.2.4",
    "pydantic-ai>=1.0.9",
]
